data/historical.py
from typing import Dict, Any, List
from datetime import datetime, timedelta
from decimal import Decimal
import pandas as pd
import numpy as np
from django.conf import settings
from dataclasses import dataclass

@dataclass
class HistoricalMetrics:
    """Data class for historical metrics"""
    avg_apr_7d: float
    volume_trend: Dict[str, Any]
    impermanent_loss_7d: float
    price_volatility: float
    key_levels: Dict[str, Decimal]

class HistoricalDataService:
    def __init__(self):
        self.time_windows = {
            '24h': timedelta(days=1),
            '7d': timedelta(days=7),
            '30d': timedelta(days=30)
        }
        
    async def get_pool_history(self, pool_address: str) -> HistoricalMetrics:
        """Get historical pool performance data"""
        try:
            # For now, return mock data
            return HistoricalMetrics(
                avg_apr_7d=25.5,
                volume_trend={
                    'direction': 'up',
                    'change_percent': 5.5,
                    'current_ma': 100000.0,
                    'week_ago_ma': 95000.0
                },
                impermanent_loss_7d=0.02,
                price_volatility=0.015,
                key_levels={
                    'strong_support': Decimal('1.05'),
                    'strong_resistance': Decimal('1.15'),
                    'support_levels': [Decimal('1.05'), Decimal('1.03'), Decimal('1.01')],
                    'resistance_levels': [Decimal('1.15'), Decimal('1.17'), Decimal('1.20')]
                }
            )
            
        except Exception as e:
            raise Exception(f"Error fetching historical data: {str(e)}") 