requirements.txt
django
djangorestframework
python-dotenv
dj-database-url
psycopg2-binary
whitenoise
gunicorn 
asgiref
aiohttp
uvicorn
channels
git+https://github.com/openai/swarm.git
django-cors-headers
numpy
web3